refactor: Piece 를 추상 클래스로 변경
