--- src/main/java/janggi/domain/Piece.java
@@ -1,20 +1,0 @@
-package janggi.domain;
-
-public class Piece {
-
-    private final Side side;
-    private final Position position;
-
-    public Piece(Side side, Position position) {
-        this.side = side;
-        this.position = position;
-    }
-
-    public Side getSide() {
-        return side;
-    }
-
-    public Position getPosition() {
-        return position;
-    }
-}
