+++ src/test/java/janggi/domain/piece/PieceTest.java
@@ -1,7 +1,10 @@
-package janggi.domain;
+package janggi.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import janggi.domain.Position;
+import janggi.domain.ReplaceUnderBar;
+import janggi.domain.Side;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
@@ -15,14 +18,26 @@ class PieceTest {
     @ParameterizedTest
     @EnumSource(value = Side.class)
     void 진영을_가진다(Side side) {
-        Piece piece = new Piece(side, DEFAULT_POSITION);
+        Piece piece = new RawPiece(side, DEFAULT_POSITION);
         assertThat(piece.getSide()).isEqualTo(side);
     }
 
     @Test
     void 위치를_가진다() {
-        Piece piece = new Piece(DEFAULT_SIDE, new Position(1, 2));
+        Piece piece = new RawPiece(DEFAULT_SIDE, new Position(1, 2));
 
         assertThat(piece.getPosition()).isEqualTo(new Position(1, 2));
+    }
+
+    private static class RawPiece extends Piece {
+
+        private RawPiece(Side side, Position position) {
+            super(side, position);
+        }
+
+        @Override
+        protected void validateMovable(int x, int y) {
+            throw new IllegalStateException("이 메서드는 테스트할 수 없습니다.");
+        }
     }
 }
